ray_gun and helix_gun to particle_gun
