+++ tests/common/include/tests/common/tools/ray_scan_utils.hpp
@@ -17,7 +17,7 @@
 #include <utility>
 #include <vector>
 
-#include "tests/common/tools/particle_gun.hpp"
+//#include "tests/common/tools/particle_gun.hpp"
 
 namespace detray {
