+++ tests/common/include/tests/common/tools_helix_gun.inl
@@ -9,7 +9,7 @@
 
 #include "detray/definitions/units.hpp"
 #include "detray/propagator/track.hpp"
-#include "tests/common/tools/helix_gun.hpp"
+#include "tests/common/tools/particle_gun.hpp"
 
 /// @note __plugin has to be defined with a preprocessor command
 using namespace detray;
@@ -18,7 +18,7 @@ using namespace __plugin;
 constexpr const float epsilon = 1e-5;
 
 // This tests the base functionality of the Helix Gun
-TEST(tools, helix_gun) {
+TEST(tools, helix_trajectory) {
 
     using vector3 = vector3<scalar>;
     using point3 = point3<scalar>;
@@ -39,29 +39,29 @@ TEST(tools, helix_gun) {
     scalar pz_along = vector::dot(mom, vector::normalize(B));
     scalar pt = std::sqrt(std::pow(p_mag, 2) - std::pow(pz_along, 2));
 
-    // helix gun
-    helix_gun helix(vertex, &B);
+    // helix trajectory
+    helix helix_traj(vertex, &B);
 
     // radius of helix
-    scalar R = helix.radius();
+    scalar R = helix_traj.radius();
     EXPECT_FLOAT_EQ(R, pt / B_mag);
 
     // After half turn
-    point3 half_turn = helix(p_mag / B_mag * M_PI);
+    point3 half_turn = helix_traj(p_mag / B_mag * M_PI);
 
     EXPECT_NEAR(half_turn[0], 0., R * epsilon);
     EXPECT_NEAR(half_turn[1], 2 * R, R * epsilon);
     EXPECT_NEAR(half_turn[2], pz_along / B_mag * M_PI, R * epsilon);
 
     // After one full turn
-    point3 full_turn = helix(2 * p_mag / B_mag * M_PI);
+    point3 full_turn = helix_traj(2 * p_mag / B_mag * M_PI);
 
     EXPECT_NEAR(full_turn[0], 0., R * epsilon);
     EXPECT_NEAR(full_turn[1], 0., R * epsilon);
     EXPECT_NEAR(full_turn[2], 2 * pz_along / B_mag * M_PI, R * epsilon);
 }
 
-TEST(tools, helix_gun_small_pT) {
+TEST(tools, helix_trajectory_small_pT) {
 
     using vector3 = vector3<scalar>;
     using point3 = point3<scalar>;
@@ -77,12 +77,12 @@ TEST(tools, helix_gun_small_pT) {
     // magnetic field
     vector3 B{0, 0, 1 * unit_constants::T};
 
-    // helix gun
-    helix_gun helix(vertex, &B);
+    // helix trajectory
+    helix helix_traj(vertex, &B);
 
     // After 10 mm
     scalar path_length = 10;
-    point3 helix_pos = helix(path_length);
+    point3 helix_pos = helix_traj(path_length);
     point3 true_pos = pos + path_length * vector::normalize(mom);
 
     EXPECT_FLOAT_EQ(true_pos[0], helix_pos[0]);
